refactor: Isolate JavaScript lockfile resolution
